pyproject.toml
[project]
name = "newsletter-agent"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "anthropic>=0.40.0",
    "feedparser>=6.0.0",
    "pydantic>=2.0.0",
    "resend>=2.0.0",
    "requests>=2.32.0",
    "readability-lxml>=0.8.1",
    "markdown>=3.7",
    "jinja2>=3.1.0",
    "newsapi-python>=0.2.7",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agent", "db", "cli"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
